+++ scilab/modules/ast/src/cpp/operations/types_comparison_eq.cpp
@@ -2,6 +2,7 @@
  *  Scilab ( http://www.scilab.org/ ) - This file is part of Scilab
  *  Copyright (C) 2012 - Scilab Enterprises - Antoine ELIAS
  *  Copyright (C) 2015 - Scilab Enterprises - Sylvain GENIN
+ *  Copyright (C) 2016 - Scilab Enterprises - Pierre-Aimé AGNEL
  *
  * Copyright (C) 2012 - 2016 - Scilab Enterprises
  *
@@ -2252,13 +2253,26 @@ InternalType* compequal_M_IC(T *_pL, U *_pR)
 template<class T, class U, class O>
 InternalType* compequal_M_E(T *_pL, U *_pR)
 {
-    return new Bool(false);
+    // Try to find an overload and call it if exists
+    types::typed_list tmp;
+    tmp.push_back(_pL);
+    tmp.push_back(_pR);
+    std::wstring overloadName(Overload::buildOverloadName(Overload::getNameFromOper(ast::OpExp::eq), tmp, 1, true));
+    types::InternalType* pIT = symbol::Context::getInstance()->get(symbol::Symbol(overloadName));
+    if (pIT)
+    {
+        return NULL;
+    }
+    else
+    {
+        return new Bool(false);
+    }
 }
 
 template<class T, class U, class O>
 InternalType* compequal_E_M(T *_pL, U *_pR)
 {
-    return new Bool(false);
+    return compequal_M_E<U, T, O>(_pR, _pL);
 }
 
 template<class T, class U, class O>