+++ scilab/CHANGES.md
@@ -564,4 +564,5 @@ Bug Fixes
 * [#14517](http://bugzilla.scilab.org/show_bug.cgi?id=14517): The second argument of part function accepted an index of 0 without exiting in error.
 * [#14524](http://bugzilla.scilab.org/show_bug.cgi?id=14524): Numeric locales were not set to standard "C" by default at scilab startup
 * [#14540](http://bugzilla.scilab.org/show_bug.cgi?id=14540): Datatips did not clip outside axes bounds
+* [#14543](http://bugzilla.scilab.org/show_bug.cgi?id=14543): fixed - == and <> operations were not overloaded for rational types
 * [#14685](http://bugzilla.scilab.org/show_bug.cgi?id=14685): datavec produced an invalid index error.